Participating Combobox is inserted
Users will see the events they are participating as checked and others as unchecked.

TODO: when combobox is selected for one event add user to participant lise of that event.

TODO: when combobox is unchecked delete that user from that event completely.

TODO: when an event is full do not allow users to participate. Can be done with a check when pressed participate, or disable the button if event is full.