Progress Bars created
Now the progress can be seen with a nice progress bar.

TextViews are stil kept, in the future will be deleted completely.

TODO: current of all events should be changed when user join or leave an event from the all events menu.

TODO: participant list should only include the ones that are true, or they should be deleted completely when they are not joining anymore.

TODO: graphical changes for the listviews